indent-common: validateCore(): Represent indents as tab/space pairs
This is a first step in being able to support smart-tabs (tabs for
indentation, then spaces for alignment); as it will need to store the 2
numbers separately.  This commit just touches the core; validateCore() and
down.  The pieces above validateCore that determine acceptable/expected
indent levels are left for later work.

This commit very liberally inserts assert() statements on the type of
arguments--since we're changing the types of a few arguments that get
passed through the system, it's very easy to miss a place and have the
wrong type somewhere.

Additionally, this includes the following changes in user-facing error
messages:

 - Say "Mixed spaces and tabs" if that's what the problem is
   * This still leaves the possibility of other whitespace characters
     showing up (like NBSP), so this doesn't totally replace the old
     "Expected {{expected}} character, but found {{actual}} character."
      message.
   * As a minor point, the message says "spaces and tabs", rather than
     "tabs and spaces", for consistency with the no-mixed-spaces-and-tabs
     rule.
 - If an line's indentation fits a style, just not the prescribed style,
   say what the found indent was.  "Expected indentation of 1 tab but found
   2 spaces.", instead of the fairly opaque unexpected character message.
 - Have "0" be unitless.  Don't say "0 tabs" or "0 spaces".  Zero is 0.