indent-common: Unify between tabs & spaces the logic for finding alignment
Reading it, wasn't immediately clear why the tabs logic worked the way it
did.  By adding it as a special case inside the the spaces logic (which is
clearer), we can make everything a little clearer.

The expression is (in pseudo-code)

    aligned_indent = max(0, expected_indent - actual_indent + columns)

But to unify it with the tabs bit, I need to pull part of it out in to a
separate variable

    extra = columns - actual_indent
    aligned_indent = max(0, expected_indent + extra)

Right away, this already is a little clearer; we can see what the
expression *means*.  However, the max() now seems a little silly, if either
expected_indent is ever negative, or if extra is ever negative, something
is *wrong*.  Drop the max(), and add assert()s that they are both positive.

    extra = columns - actual_indent
    assert(extra > 0)
    assert(expected_indent > 0)
    aligned_indent = max(0, expected_indent + extra)